fixed the way characters are loaded (hopefully)
previously characters where fetched using a single function with async/await, which caused each character to load seperately, often leading to longer load times. This change will save all characters to an array, and when this array matches the amount of characters, they are then loaded in. This reduces the character cards load time by a lot.